RavenDB-18645 - GET /databases/*/studio/collections/fields
